Refector many things, add ability to change shape with animation, update gitignore.

+++ BeatShape/Assets/Beatboard/beatboardData.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class beatboardData : MonoBehaviour
 {
-    public int points;
+    public float points;
     public float size;
 }
